Refactor around transient
