+++ tests/includes/test-class-static-press-repository.php
@@ -9,8 +9,8 @@ namespace static_press\tests\includes;
 
 require_once dirname( __FILE__ ) . '/../testlibraries/class-expect-url.php';
 require_once dirname( __FILE__ ) . '/../testlibraries/class-model-url.php';
-use static_press\includes\Static_Press_Ajax_Init;
 use static_press\includes\Static_Press_Repository;
+use static_press\includes\Static_Press_Transient_Service;
 use static_press\includes\Static_Press_Url_Updater;
 use static_press\tests\testlibraries\Expect_Url;
 use static_press\tests\testlibraries\Model_Url;
@@ -34,20 +34,12 @@ class Static_Press_Repository_Test extends \WP_UnitTestCase {
 			),
 		);
 
-		$repository   = new Static_Press_Repository();
-		$static_press = new Static_Press_Ajax_Init(
-			null,
-			null,
-			$repository,
-			null
-		);
-		$url_updater  = new Static_Press_Url_Updater( $repository, null );
+		$repository  = new Static_Press_Repository();
+		$url_updater = new Static_Press_Url_Updater( $repository, null );
 		$url_updater->update( $urls );
-		$reflection = new \ReflectionClass( get_class( $static_press ) );
-		$method     = $reflection->getMethod( 'fetch_start_time' );
-		$method->setAccessible( true );
-		$start_time = $method->invokeArgs( $static_press, array() );
-		$results    = $repository->get_all_url( $start_time );
+		$transient_service = new Static_Press_Transient_Service();
+		$start_time        = $transient_service->fetch_start_time();
+		$results           = $repository->get_all_url( $start_time );
 		Expect_Url::assert_url(
 			$this,
 			array(
